+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the `math` library to the Standard Library
 - Added the `os` library to the Standard Library
 - Updated the `string` library to add new features
+- Added error handling via [codespan_reporting](https://crates.io/crates/codespan-reporting)
 
 ## 0.1.0 (2025-04-17)
 - Added the `string` library in the Mussel Standard Library to manipulate strings